rename abstract-core re-export in sdk
