+++ framework/packages/abstract-sdk/src/ans_resolve.rs
@@ -1,7 +1,7 @@
 //! # AnsHost Entry
 //! An entry (value) in the ans_host key-value store.
 
-use crate::core::objects::{
+use crate::framework::objects::{
     ans_host::AnsHost, pool_metadata::ResolvedPoolMetadata, AnsAsset, AssetEntry, ChannelEntry,
     ContractEntry, DexAssetPairing, LpToken, PoolMetadata, PoolReference, UniquePoolId,
 };
@@ -369,7 +369,7 @@ mod tests {
     mod pool_metadata {
         use super::*;
 
-        use crate::core::objects::PoolType;
+        use crate::framework::objects::PoolType;
 
         #[test]
         fn exists() {
@@ -425,7 +425,7 @@ mod tests {
 
     mod pools {
         use super::*;
-        use crate::core::objects::{PoolAddress, PoolType};
+        use crate::framework::objects::{PoolAddress, PoolType};
         use abstract_core::ans_host::state::{ASSET_PAIRINGS, POOL_METADATA};
 
         #[test]
@@ -474,7 +474,7 @@ mod tests {
 
     mod contract_entry {
         use super::*;
-        use crate::core::ans_host::state::CONTRACT_ADDRESSES;
+        use crate::framework::ans_host::state::CONTRACT_ADDRESSES;
 
         #[test]
         fn exists() {
@@ -549,7 +549,7 @@ mod tests {
         use abstract_core::objects::chain_name::ChainName;
 
         use super::*;
-        use crate::core::ans_host::state::CHANNELS;
+        use crate::framework::ans_host::state::CHANNELS;
         use std::str::FromStr;
 
         #[test]
@@ -586,7 +586,7 @@ mod tests {
 
     mod asset_info_and_asset {
         use super::*;
-        use crate::core::ans_host::state::REV_ASSET_ADDRESSES;
+        use crate::framework::ans_host::state::REV_ASSET_ADDRESSES;
 
         #[test]
         fn exists() {
