+++ framework/packages/abstract-sdk/src/apis/adapter.rs
@@ -110,7 +110,7 @@ mod tests {
     }
     mod adapter_request {
         use super::*;
-        use crate::core::adapter::{self, AdapterRequestMsg};
+        use crate::framework::adapter::{self, AdapterRequestMsg};
 
         #[test]
         fn should_return_err_if_not_dependency() {
