+++ framework/packages/abstract-sdk/src/base/features/dependencies.rs
@@ -1,5 +1,5 @@
 use crate::base::Handler;
-use crate::core::objects::dependency::StaticDependency;
+use crate::framework::objects::dependency::StaticDependency;
 
 /// Retrieve the dependencies of a module.
 pub trait Dependencies: Sized {
