+++ framework/packages/abstract-sdk/src/base/features/identification.rs
@@ -1,4 +1,4 @@
-use crate::core::objects::AccountId;
+use crate::framework::objects::AccountId;
 use abstract_core::{
     objects::common_namespace::ADMIN_NAMESPACE, proxy::state::ACCOUNT_ID,
     version_control::AccountBase,
