+++ framework/packages/abstract-sdk/src/mock_module.rs
@@ -4,7 +4,7 @@ use abstract_core::objects::dependency::StaticDependency;
 use abstract_testing::prelude::{TEST_MODULE_ID, TEST_PROXY};
 use cosmwasm_std::{Addr, Deps};
 
-use crate::core::objects::module::ModuleId;
+use crate::framework::objects::module::ModuleId;
 use crate::features::{
     AbstractNameService, AbstractRegistryAccess, AccountIdentification, Dependencies,
     ModuleIdentification,
